SIP: use expiry time from Contact header
RFC3261 section 10.3 does not mention the "Expires" header,
but does say the Contact headers MUST be present, and that
the expires parameter in those MUST be present.

At least one registrar (sipgate.com) does not provide an
Expires header in the response, but does provide the expires
parameter in the appropriate Contact header.