libstagefright: fix live RTSP streaming on some servers
'npt=0-' is also a valid range to describe an "infinite" session, and
it's used by some popular RTSP servers in mobile environments (Envivio's,
for example). Expecting only 'now-' was causing these sessions to fail (and
thus breaking the "Mobile TV" service).

Also, drop RTP video packets that are out of sequence. Trying to
reorder these was causing a *very* long time in starting the streams.