Fix a race condition in NativeDaemonConnector
This fixes the issue where with a BFS kernel SD cards may fail to mount on boot.

Fixes a race between the onDaemonConnected callback and setting the mOutputStream in NativeDaemonConnector.

MountService connects to vold using the NativeDaemonConnector. Throws a "NativeDaemonConnectorException: No output stream!" when the onDaemonConnected callback in MountService calls the doListCommand.